allow kore-exec to use either deserialized or KORE definition
