+++ kore/kore.cabal
@@ -130,8 +130,13 @@ common global-main
     Paths_kore
   hs-source-dirs: app/share
   build-depends: kore
+  build-depends: binary >=0.8.8.0
+  build-depends: bytestring >=0.10
   build-depends: clock >=0.8
+  build-depends: compact >= 0.2.0.0
   build-depends: containers >=0.5
+  build-depends: deepseq >=1.4
+  build-depends: exceptions >=0.10
   build-depends: filepath >=1.4
   build-depends: generic-lens >=1.1
   build-depends: lens >=4.17
@@ -148,7 +153,6 @@ common library
   build-depends: clock >=0.8
   build-depends: co-log >=0.4
   build-depends: comonad >=5.0
-  build-depends: compact >= 0.2.0.0
   build-depends: containers >=0.5.8
   build-depends: cryptonite >=0.25
   build-depends: data-default >=0.7
@@ -586,7 +590,6 @@ executable kore-exec
   hs-source-dirs: app/exec
   build-depends: compact >=0.2.0.0
   build-depends: data-default >=0.7
-  build-depends: deepseq >=1.4
   build-depends: directory >=1.3
   build-depends: exceptions >=0.10
   build-depends: extra >=1.6
